Ignore whitespace in user input
Due to way the chunks are shown with space between them it's not completely
clear whether the input should also contain whitespace or not. A solution is
to just strip all whitespace from the input. The problem is that, if the word
being memorized contains whitespace, we will ignore it also.

A better solution will most likely be needed.