Handle whitespace in user input better
Previously we ingored whitespace in user input because the spaces in the
screen were seen as confusing. Problem was, this made memorizing words with
spaces in them not work properly.
To solve this, we underline the text so that spaces are easier to recognize
and no longer ignore whitespace in user input.