Revamp sequential mode
Chunks are now shown one at a time. When quessing, already seen chunks
are displayed on the left. Also, since we only show one chunk at a time, we show it
in the middle instead of having left and right sides like in normal mode.
All of this requires some checks in the different views to do different things based
on the mode we are in. Having to check the mode in bunch of places seems a bit hacky
but works for now.