Redesign visualization
Completely overhauls the UI and changes the behaviour somewhat.
Instead of showing only one node of the chunk tree at a time we show
the whole tree, to better help visualize the chunks. We still do a post-order
traversal of the tree, and when the input is wrong we go back in the tree.