Call done() last
Main reason for this change is that we want to increment memorisedCount
only after the log data is sent. This way memorisedCount is the same
for all guesses until we restart from the beginning, instead of the
completing guess having a different memorisedCount.
This will make querying the data in the backend easier.